Add RS485 battery and inverter protocol base classes and refactor RS485/Modbus handling
