+++ Software/src/inverter/BYD-MODBUS.h
@@ -5,16 +5,22 @@
 #define MODBUS_INVERTER_SELECTED
 #define SELECTED_INVERTER_CLASS BydModbusInverter
 
-#define MB_RTU_NUM_VALUES 13100
-#define MAX_POWER 40960  //BYD Modbus specific value
-
-extern uint16_t mbPV[MB_RTU_NUM_VALUES];
+#include "ModbusInverterProtocol.h"
 
 class BydModbusInverter : public ModbusInverterProtocol {
  public:
   void setup();
-  void update_modbus_registers();
+  void update_values();
+
+ private:
   void handle_static_data();
+  void verify_temperature();
+  void verify_inverter_modbus();
+  void handle_update_data_modbusp201_byd();
+  void handle_update_data_modbusp301_byd();
+
+  //BYD Modbus specific value
+  const int MAX_POWER = 40960;
 };
 
 #endif
