+++ Software/src/inverter/ModbusInverterProtocol.h
@@ -1,0 +1,23 @@
+#ifndef MODBUS_INVERTER_PROTOCOL_H
+#define MODBUS_INVERTER_PROTOCOL_H
+
+#include <stdint.h>
+#include "../lib/eModbus-eModbus/ModbusServerRTU.h"
+#include "HardwareSerial.h"
+#include "InverterProtocol.h"
+
+extern uint16_t mbPV[];
+
+// The abstract base class for all Modbus inverter protocols
+class ModbusInverterProtocol : public InverterProtocol {
+ protected:
+  // Create a ModbusRTU server instance with 2000ms timeout
+  ModbusInverterProtocol() : MBserver(2000) { mbPV = ::mbPV; }
+
+  static const int MB_RTU_NUM_VALUES = 13100;
+  uint16_t* mbPV;  // Process variable memory
+
+  ModbusServerRTU MBserver;
+};
+
+#endif
